v4 fix docstring (#575)
* wordWrap break at first place that NEXT position over the line limit

* regen tests
