+++ fluentnamer/readme.md
@@ -6,7 +6,7 @@ pass-thru:
   - subset-reducer
 
 use-extension:
-  "@autorest/modelerfour": "4.12.304"
+  "@autorest/modelerfour": "4.13.309"
 
 pipeline:
